requirements.txt
click
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
dnspython==1.16.0
idna==2.8
lxml==4.3.2
numpy==1.16.1
pandas==0.24.1
pyactiveresource==2.2.1
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==5.3.1
requests==2.21.0
ShopifyAPI==8.0.2
six==1.12.0
twill3==0.11.2
Unidecode==1.0.23
urllib3==1.24.1
